+++ src/test/java/org/springframework/samples/petclinic/owner/PetControllerTests.java
@@ -67,7 +67,7 @@ class PetControllerTests {
 	@BeforeEach
 	void setup() {
 		PetType cat = new PetType();
-		cat.setId(3);
+		cat.setId(3L);
 		cat.setName("hamster");
 		given(this.types.findPetTypes()).willReturn(List.of(cat));
 
@@ -76,8 +76,8 @@ class PetControllerTests {
 		Pet dog = new Pet();
 		owner.addPet(pet);
 		owner.addPet(dog);
-		pet.setId(TEST_PET_ID);
-		dog.setId(TEST_PET_ID + 1);
+		pet.setId((long) TEST_PET_ID);
+		dog.setId((long) (TEST_PET_ID + 1));
 		pet.setName("petty");
 		dog.setName("doggy");
 		given(this.owners.findById(TEST_OWNER_ID)).willReturn(Optional.of(owner));
