+++ src/test/java/org/springframework/samples/petclinic/owner/VisitControllerTests.java
@@ -60,7 +60,7 @@ class VisitControllerTests {
 		Owner owner = new Owner();
 		Pet pet = new Pet();
 		owner.addPet(pet);
-		pet.setId(TEST_PET_ID);
+		pet.setId((long) TEST_PET_ID);
 		given(this.owners.findById(TEST_OWNER_ID)).willReturn(Optional.of(owner));
 	}
